🤖 refactor: shelter mypage event history 컴포넌트 분리
